requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
cffi==1.17.1
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.58.0
greenlet==3.2.2
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
pefile==2023.2.7
pillow==11.2.1
pip==25.1.1
psutil==7.0.0
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyNaCl==1.5.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
scipy==1.15.3
setuptools==80.4.0
sip==6.10.0
six==1.17.0
soundfile==0.13.1
tzdata==2025.2
urllib3==2.4.0
